requirements.txt
Django==2.2.6
python-dateutil==2.8.0
